Restructure: move GAE kernel to rl_games/triton_kernels/gae_kernel.py
Move both PyTorch and Triton implementations into gae_kernel.py with
the dispatch logic (USE_TRITON switch) inside compute_gae(). The caller
(a2c_common.py) now just calls compute_gae() without knowing about the
backend.
