+++ rl_games/triton_kernels/__init__.py
@@ -1,0 +1,1 @@
+from rl_games.triton_kernels.gae_kernel import compute_gae
